Refactor alarms_parser dependency on dita_content
