+++ metaswitch/common/alarms_parser.py
@@ -34,7 +34,6 @@ import json
 import StringIO
 import csv
 import alarm_severities
-from dita_content import DITAContent
 
 # Valid severity levels - this should be kept in sync with the
 # list in alarmdefinition.h in cpp-common
@@ -259,46 +258,6 @@ def alarms_to_csv(alarms_files):
                 writer.writerow(values)
 
     return output.getvalue()
-
-
-# Read in alarm information from a list of alarms files and generate a DITA
-# document describing the alarms.   Returns DITA as XML.
-def alarms_to_dita(alarms_files):
-    dita_content = DITAContent()
-    dita_content.begin_section("Alarms")
-
-    for alarm_file in alarms_files:
-        alarm_list = parse_alarms_file(alarm_file)
-
-        for alarm in alarm_list:
-            for alarm_level in alarm._levels.itervalues():
-                fields = {"OID": full_alarm_oid(alarm_level._oid),
-                          "ITU severity": alarm_level._itu_severity,
-                          "Severity": alarm_level._severity_string,
-                          "Description": alarm_level._description,
-                          "Details": alarm_level._details,
-                          "Cause": alarm_level._cause,
-                          "Effect": alarm_level._effect,
-                          "Action": alarm_level._action}
-
-                dita_content.begin_table(alarm._name + ": " + alarm_level._severity_string,
-                                         ["Field", "Value"],
-                                         ["25%", "75%"])
-                for field, value in fields.iteritems():
-                    dita_content.add_table_entry([field, value])
-                dita_content.end_table()
-
-    dita_content.end_section()
-    return dita_content._xml
-
-
-# Read in alarm information from a list of alarms files and write a DITA
-# document describing them.
-def write_dita_file(alarms_files, dita_filename): #pragma: no cover
-    xml = alarms_to_dita(alarms_files)
-
-    with open(dita_filename, "w") as dita_file:
-        dita_file.write(xml)
 
 
 # Read in alarm information from a list of alarms files and write a CSV
